Replace spent_at with block.slot_no in outer join
When joining the transactions with the block that they are supposed to
have been spent in, we correctly set header hash to null if such a block
is not in the checkpoints anymore. However, the spent_at tag is not set
to null. Therefore we run into the possibility that we mark a result
as spent (spent_at is not null) even though it is not (the spent_at
block does not exist in checkpoints anymore).